move Task classes to task package
