+++ music-universe/music-data-raw-lastfm/src/main/java/yurykorzun/art/universe/music/data/raw/lastfm/task/service/LastfmTaskService.java
@@ -1,6 +1,6 @@
 package yurykorzun.art.universe.music.data.raw.lastfm.task.service;
 
-import yurykorzun.art.universe.common.persistence.entity.TaskStatus;
+import yurykorzun.art.universe.common.data.raw.task.entity.TaskStatus;
 import yurykorzun.art.universe.music.data.raw.lastfm.task.dto.LastfmTaskCreateRequest;
 
 public interface LastfmTaskService {
